Анимация материала независимо по трём каналам цвета

+++ viewer/include/animations-list.h
@@ -8,7 +8,6 @@
 #include <vsg/core/Object.h>
 #include <vsg/core/ref_ptr.h>
 
-#include <cstddef>
 #include <map>
 #include <mutex>
 #include <utility>
@@ -19,9 +18,9 @@
 struct animations_t final : public vsg::Inherit<vsg::Object, animations_t>
 {
     std::mutex mutex;
-    std::multimap<std::size_t, vsg::ref_ptr<ProcAnimation>> animations;
+    std::multimap<std::int32_t, vsg::ref_ptr<ProcAnimation>> animations;
 
-    void thread_safe_insert(std::pair<std::size_t, vsg::ref_ptr<ProcAnimation>> id_and_animation);
+    void thread_safe_insert(std::pair<std::int32_t, vsg::ref_ptr<ProcAnimation>> id_and_animation);
 };
 
 #endif // ANIMATIONS_LIST_H
